32 Fix Use of undefined constant CONTEXT_COHORT
Previously deleting a cohort or group the error "Use of undefined
constant CONTEXT_COHORT" or "Use of undefined constant CONTEXT_GROUP"
would occur.  This change also removes the require_once() from the
PHPUnit test as this brought these constants into scope causing the test
to pass when it should have failed (because of this problem).